+++ R/splat-estimate.R
@@ -277,8 +277,16 @@ splatEstDropout <- function(norm.counts, params) {
     error = function(err) {
         warning("Fitting dropout using the Gauss-Newton method failed, ",
                 "using the Golub-Pereyra algorithm instead")
-        nls(y ~ logistic(x, x0 = x0, k = k), data = df,
-            start = list(x0 = 0, k = -1), algorithm = "plinear")
+        tryCatch({
+            nls(y ~ logistic(x, x0 = x0, k = k), data = df,
+                start = list(x0 = 0, k = -1), algorithm = "plinear")
+        },
+        error = function(err) {
+            warning("Fitting dropout using the Golub-Pereyra method failed, ",
+                    "using the nl2sol algorithm instead")
+            nls(y ~ logistic(x, x0 = x0, k = k), data = df,
+                start = list(x0 = 0, k = -1), algorithm = "port")
+        })
     })
 
     #exp.zeros <- dnbinom(0, mu = means, size = 1 / 0.1) * ncol(norm.counts)